Separate util func
